Move Statement and Rows to separate file
